Move sorted faces rendering to RendererSorted.cpp
