+++ TombEngine/Renderer/RendererSprites.cpp
@@ -418,118 +418,6 @@ namespace TEN::Renderer
 		_graphicsDevice->SetPrimitiveType(PrimitiveType::TriangleList);
 	}
 
-	void Renderer::DrawSingleSprite(RendererSortableObject* object, RendererObjectType lastObjectType, RenderView& view)
-	{
-		_graphicsDevice->SetPrimitiveType(PrimitiveType::TriangleStrip);
-
-		BindRenderTargetAsTexture(TextureRegister::GBufferDepthMap, _depthRenderTarget->GetRenderTarget(), SamplerStateRegister::LinearClamp);
-
-		SetDepthState(DepthState::Read);
-		SetCullMode(CullMode::None);
-		SetBlendMode(object->Sprite->BlendMode);
-		SetAlphaTest(AlphaTestMode::GreatherThan, ALPHA_TEST_THRESHOLD);
-
-		_shaders.Bind(Shader::InstancedSprites);
-
-		_stInstancedSpriteBuffer.Sprites[0].World = object->Sprite->Type != SpriteType::ThreeD ?
-			GetWorldMatrixForSprite(*object->Sprite, view) :
-			Matrix::Identity;
-		_stInstancedSpriteBuffer.Sprites[0].PerVertexColor = 1;
-		_stInstancedSpriteBuffer.Sprites[0].IsSoftParticle = object->Sprite->SoftParticle ? 1 : 0;
-		_stInstancedSpriteBuffer.Sprites[0].RenderType = (int)object->Sprite->renderType;
-
-		PackSpriteTextureCoordinates(0, object->Sprite->Sprite);
-
-		UpdateConstantBuffer(&_stInstancedSpriteBuffer, _cbInstancedSpriteBuffer.get());;
-
-		BindTexture(TextureRegister::ColorMap, object->Sprite->Sprite->Texture, SamplerStateRegister::LinearClamp);
-		
-		// Set up vertex buffer and parameters.
-		if (object->Sprite->Type != SpriteType::ThreeD)
-		{
-			_graphicsDevice->BindVertexBuffer(_quadVertexBuffer.get());
-		}
-		else
-		{
-			auto vertex0 = Vertex{};
-			vertex0.Position = object->Sprite->vtx1;
-			vertex0.UV = object->Sprite->Sprite->UV[0];
-			vertex0.Color = VectorColorToRGBA(object->Sprite->c1);
-			vertex0.Effects = 0 << INDEX_IN_POLY_VERTEX_SHIFT;
-
-			auto vertex1 = Vertex{};
-			vertex1.Position = object->Sprite->vtx2;
-			vertex1.UV = object->Sprite->Sprite->UV[1];
-			vertex1.Color = VectorColorToRGBA(object->Sprite->c2);
-			vertex1.Effects = 1 << INDEX_IN_POLY_VERTEX_SHIFT;
-
-			auto vertex2 = Vertex{};
-			vertex2.Position = object->Sprite->vtx3;
-			vertex2.UV = object->Sprite->Sprite->UV[2];
-			vertex2.Color = VectorColorToRGBA(object->Sprite->c3);
-			vertex2.Effects = 2 << INDEX_IN_POLY_VERTEX_SHIFT;
-
-			auto vertex3 = Vertex{};
-			vertex3.Position = object->Sprite->vtx4;
-			vertex3.UV = object->Sprite->Sprite->UV[3];
-			vertex3.Color = VectorColorToRGBA(object->Sprite->c4);
-			vertex3.Effects = 3 << INDEX_IN_POLY_VERTEX_SHIFT;
-
-			_spriteVertices.clear();
-			_spriteVertices.push_back(vertex0);
-			_spriteVertices.push_back(vertex1);
-			_spriteVertices.push_back(vertex3);
-			_spriteVertices.push_back(vertex2);
-
-			_graphicsDevice->UpdateVertexBuffer(_spriteVertexBuffer.get(), 0, 4, _spriteVertices.data());
-			_graphicsDevice->BindVertexBuffer(_spriteVertexBuffer.get());
-		}
-
-		// Draw sprites with instancing.
-		DrawInstancedTriangles(4, 1, 0);
-
-		_numSortedSpritesDrawCalls++;
-		_numSortedTriangles += 2;
-
-		_graphicsDevice->SetPrimitiveType(PrimitiveType::TriangleList);
-	}
-
-	void Renderer::DrawSpriteSorted(RendererSortableObject* objectInfo, RendererObjectType lastObjectType, RenderView& view, int baseVertex, int count)
-	{
-		if (lastObjectType != objectInfo->ObjectType)
-		{
-			_shaders.Bind(Shader::InstancedSprites);
-
-			_graphicsDevice->BindVertexBuffer(_sortedPolygonsVertexBuffer.get());
-			_graphicsDevice->SetPrimitiveType(PrimitiveType::TriangleList);
-			_graphicsDevice->SetInputLayout(_vertexInputLayout.get());
-
-		}
-
-		_stInstancedSpriteBuffer.Sprites[0].World = Matrix::Identity;
-		_stInstancedSpriteBuffer.Sprites[0].PerVertexColor = 1;
-		_stInstancedSpriteBuffer.Sprites[0].IsSoftParticle = objectInfo->Sprite->SoftParticle ? 1 : 0;
-		_stInstancedSpriteBuffer.Sprites[0].RenderType = (int)objectInfo->Sprite->renderType;
-
-		PackSpriteTextureCoordinates(0, objectInfo->Sprite->Sprite);
-
-		// Only Sprites[0] is used, so upload just the CB prefix.
-		UpdateConstantBuffer(&_stInstancedSpriteBuffer, _cbInstancedSpriteBuffer.get(), (int)sizeof(InstancedSprite));
-
-		SetDepthState(DepthState::Read);
-		SetCullMode(CullMode::None);
-		SetBlendMode(objectInfo->Sprite->BlendMode);
-		SetAlphaTest(AlphaTestMode::None, ALPHA_TEST_THRESHOLD);
-
-		BindTexture(TextureRegister::ColorMap, objectInfo->Sprite->Sprite->Texture, SamplerStateRegister::LinearClamp);
-		BindRenderTargetAsTexture(TextureRegister::GBufferDepthMap, _depthRenderTarget->GetRenderTarget(), SamplerStateRegister::PointWrap);
-
-		DrawInstancedTriangles(count, 1, baseVertex);
-
-		_numSortedSpritesDrawCalls++;
-		_numSortedTriangles += count / 3;
-	}
-
 	void Renderer::PackSpriteTextureCoordinates(int instanceId, RendererSprite* sprite)
 	{
 		// NOTE: Strange packing due to particular HLSL 16-byte alignment requirements.
